#TUS-11-T-2 Fix potential bugs #TUS-5-IUS-1-T-4 Write pdf files

+++ clinical.center/src/main/java/mrs/isa/team12/clinical/center/service/ClinicImpl.java
@@ -12,6 +12,7 @@ import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import javassist.expr.Cast;
 import mrs.isa.team12.clinical.center.dto.ClinicDto;
 import mrs.isa.team12.clinical.center.model.Appointment;
 import mrs.isa.team12.clinical.center.model.Clinic;
@@ -106,7 +107,7 @@ public class ClinicImpl implements ClinicService {
 		for(Rating r : clinicToUpdate.getRatings()) {
 			ratingSum += r.getRating();
 		}
-		double newRating = ratingSum / clinicToUpdate.getRatings().size();
+		double newRating = (double) ratingSum / clinicToUpdate.getRatings().size();
 		clinicToUpdate.setRating(newRating);
 		Clinic updated = repository.save(clinicToUpdate);
 		logger.info("< update id{}", updated.getId());
